[Task/#0070] - Modify backend codes (model & controller layer)

See merge request andrewmoplas/restaurant-ordering-system!81

+++ src/main/java/com/webtech/Service/FeedBackService.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 import com.webtech.Dao.FeedBackDao;
 import com.webtech.Dao.OrderDao;
 import com.webtech.Model.FeedBack;
+import com.webtech.Model.Menu;
 import com.webtech.Model.Order;
 import com.webtech.Model.OrderStatus;
 
@@ -17,30 +18,37 @@ public class FeedBackService implements SERVICE<FeedBack> {
 	private FeedBackDao repository;
 	
 	@Override
-	public FeedBack create(FeedBack obj) {
-		return repository.create(obj);
+	public List<FeedBack> create(FeedBack obj) {
+		repository.create(obj);
+		return repository.getItems();
 	}
 
 	@Override
-	public void update(FeedBack obj) {
+	public List<FeedBack> update(FeedBack obj) {
 		repository.update(obj);
+		return repository.getItems();
 	}
 
 	@Override
-	public boolean delete(String id) {
+	public List<FeedBack> delete(String id) {
 		boolean result = false;
 		
 		try {		
 			Long longId = Long.parseLong(id);
-			
-			
+			FeedBack menu = repository.getItem(id);
+			if(menu != null) {
 				result = repository.delete(id);
+				return repository.getItems();
+			}else {
+				return null;
+			}
 			
 		} catch (Exception ex) {
 			System.out.print("DELETESERVICE");
+			return null;
 		}
 		
-		return result;
+		
 	}
 
 	@Override
